project_7_First_AIAgent/requirements.txt
langchain
langchain_google_genai
python-dotenv
langchain_community